% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_effects.R
\name{add_variant_effect}
\alias{add_variant_effect}
\title{Add effect of a variant to an outcome}
\usage{
add_variant_effect(outcome, G, h2)
}
\arguments{
\item{outcome}{vector}

\item{G}{genotype vector}

\item{h2}{heritability}
}
\value{
outcome with variant effect
}
\description{
Add effect of a variant to an outcome
}
