% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_effects.R
\name{heritability}
\alias{heritability}
\title{Calculate heritability from beta}
\usage{
heritability(G, beta, varComp)
}
\arguments{
\item{G}{genotype vector}

\item{beta}{effect size}

\item{varComp}{2-element vector with (genetic, error) variance components}
}
\value{
h2 (heritability)
}
\description{
Calculate heritability from beta
}
