% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_pruning.R
\name{ldprune_strata}
\alias{ldprune_strata}
\title{LD prune variants in all strata}
\usage{
ldprune_strata(
  gdsobj,
  strata,
  variant.id = NULL,
  maf = 0.01,
  ld.threshold = sqrt(0.1),
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{gdsobj}{\code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{strata}{named list of sample.id in strata}

\item{variant.id}{ids of variants to consider}

\item{maf}{MAF threshold for pruning in pooled sample}

\item{ld.threshold}{LD pruning threshold: single value or named list matching \code{strata}}

\item{seed}{seed for LD pruning: single value or named list matching \code{strata}}

\item{verbose}{logical for whether to print messages}
}
\value{
vector with variant.id
}
\description{
LD prune variants in all strata
}
