% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{outcome_from_covMat}
\alias{outcome_from_covMat}
\alias{outcome_from_covMat_blocks}
\title{Generate random normal outcome from covariance matrix}
\usage{
outcome_from_covMat(covMat, varComp, outcome = NULL)

outcome_from_covMat_blocks(covMat, varComp, outcome = NULL)
}
\arguments{
\item{covMat}{a covariance matrix}

\item{varComp}{2-element vector with (genetic, error) variance components}

\item{outcome}{vector to be adjusted for covariance. If \code{NULL}, will be created using \code{\link{rnorm}}.}
}
\value{
numeric vector of outcomes
}
\description{
If \code{covMat} is a block diagonal matrix, use \code{outcome_from_covMat_blocks}.
}
