% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_effects.R
\name{power}
\alias{power}
\title{Calculate power}
\usage{
power(N, h2, pval = 5e-08)
}
\arguments{
\item{N}{number of samples}

\item{h2}{heritability}

\item{pval}{p-value threshold for significance}
}
\value{
power
}
\description{
Calculate power
}
\references{
https://github.com/kaustubhad/gwas-power/blob/master/power_calc_functions.R
}
