% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_selection.R
\name{var_multiple_strata}
\alias{var_multiple_strata}
\alias{var_single_stratum}
\title{Select variants based on allele frequency}
\usage{
var_multiple_strata(a, b, min.freq = 0.1)

var_single_stratum(a, b, min.freq = 0.1, max.freq = 0.001)
}
\arguments{
\item{a}{vector of variant frequency in group a}

\item{b}{vector of variant frequency in group b}

\item{min.freq}{miniumum frequency to be included in results}

\item{max.freq}{max frequency to be considered "not present" in group b}
}
\value{
index of variant meeting criteria
}
\description{
\code{var_multiple_strata} selects variants present
in groups a and b with frequency \code{>min.freq}
}
\details{
\code{var_single_stratum} selects variants
with frequency \code{>min.freq} in group a and
frequency \code{<max.freq} in group b
}
