% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_effects.R
\name{variant_assoc}
\alias{variant_assoc}
\title{Run association test for an outcome and a variant}
\usage{
variant_assoc(variant.sel, gdsobj, outcome, covars, cov.mat, beta, varComp)
}
\arguments{
\item{variant.sel}{index of variant in unfilted gdsobj}

\item{gdsobj}{SeqVarGDSClass object}

\item{outcome}{vector of outcome}

\item{covars}{data.frame of covariates}

\item{cov.mat}{covariance matrix}

\item{beta}{effect size for variant}

\item{varComp}{2-element vector with (genetic, error) variance components}
}
\value{
association test results for outcome and variant
}
\description{
Run association test for an outcome and a variant
}
\details{
Order of samples in \code{outcome} and \code{covars} must match \code{cov.mat}. The rownames of \code{cov.mat} must correspond to the \code{sample.id} node in \code{gdsobj}.
}
