% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_effects.R
\name{variant_assoc}
\alias{variant_assoc}
\title{Run association test for an outcome and a variant}
\usage{
variant_assoc(
  variant.sel,
  h2 = NULL,
  beta = NULL,
  varComp,
  gdsobj,
  dat,
  outcome,
  cov.mat,
  covars = NULL
)
}
\arguments{
\item{variant.sel}{index of variant in unfilted gdsobj}

\item{h2}{heritability}

\item{beta}{effect size for variant}

\item{varComp}{2-element vector with (genetic, error) variance components}

\item{gdsobj}{SeqVarGDSClass object}

\item{dat}{AnnotatedDataFrame with sample.id, outcome, and covariates}

\item{outcome}{character string specifying the name of the outcome variable in \code{dat}}

\item{cov.mat}{covariance matrix}

\item{covars}{A vector of character strings specifying the names of the fixed effect covariates in \code{dat}}
}
\value{
association test results for outcome and variant
}
\description{
Run association test for an outcome and a variant
}
\details{
Either h2 or beta must be specified.

Order of samples in \code{outcome} and \code{covars} must match \code{cov.mat}. The rownames of \code{cov.mat} must correspond to the \code{sample.id} node in \code{gdsobj}.
}
