% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_effects.R
\name{variant_genotypes}
\alias{variant_genotypes}
\title{Return genotypes for set of samples and variants}
\usage{
variant_genotypes(gdsobj, variant.id, sample.id = NULL)
}
\description{
Return genotypes for set of samples and variants
}
